Modularize Codex usage handling
